Se hace el empaquetamiento requerido de Controller y Model
